+++ parquet/parquet-hadoop/src/main/java/org/apache/parquet/hadoop/ParquetWriter.java
@@ -453,7 +453,7 @@ public class ParquetWriter<T> implements Closeable {
     }
   }
 
-  public float[] report() {
+  public long[] report() {
     return writer.report();
   }
